Make openai-whisper an actual optional dep
